ref: Move Backtrace parsing out of failure integration
In preparation to split things up further.
